# openapi.json
{"openapi":"3.1.0","info":{"title":"Émergence API","version":"7.0"},"paths":{"/api/health":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/documents/":{"get":{"tags":["Documents","Documents"],"summary":"List Documents","description":"Retourne la liste de tous les documents uploadés.","operationId":"list_documents_api_documents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Documents Api Documents  Get"}}}}}}},"/api/documents":{"get":{"tags":["Documents","Documents"],"summary":"List Documents Alias","description":"Alias sans slash final pour éviter 404.","operationId":"list_documents_alias_api_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Documents Alias Api Documents Get"}}}}}}},"/api/documents/upload":{"post":{"tags":["Documents","Documents"],"summary":"Upload Document","operationId":"upload_document_api_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_documents_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/documents/{document_id}":{"delete":{"tags":["Documents","Documents"],"summary":"Delete Document","description":"Supprime un document, ses chunks et ses vecteurs.","operationId":"delete_document_api_documents__document_id__delete","parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"string","title":"Document Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/debate/{debate_id}":{"get":{"summary":"Get Debate Details","description":"Récupère les détails d'un débat actif.","operationId":"get_debate_details_api_debate__debate_id__get","parameters":[{"name":"debate_id","in":"path","required":true,"schema":{"type":"string","title":"Debate Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/debate/":{"get":{"summary":"List Active Debates","description":"Liste tous les débats actuellement en cours d'exécution.","operationId":"list_active_debates_api_debate__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/dashboard/costs/summary":{"get":{"tags":["Dashboard"],"summary":"Récupère le résumé complet des données du cockpit","description":"Fournit un résumé des coûts, des métriques de monitoring et les seuils d'alerte.","operationId":"get_dashboard_summary_api_dashboard_costs_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Dashboard Summary Api Dashboard Costs Summary Get"}}}}}}},"/api/threads/":{"get":{"tags":["Threads","Threads"],"summary":"List Threads","operationId":"list_threads_api_threads__get","parameters":[{"name":"type","in":"query","required":false,"schema":{"anyOf":[{"type":"string","pattern":"^(chat|debate)$"},{"type":"null"}],"title":"Type"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Offset"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Threads","Threads"],"summary":"Create Thread","operationId":"create_thread_api_threads__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/threads/{thread_id}":{"get":{"tags":["Threads","Threads"],"summary":"Get Thread","operationId":"get_thread_api_threads__thread_id__get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}},{"name":"messages_limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Messages Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Threads","Threads"],"summary":"Update Thread","operationId":"update_thread_api_threads__thread_id__patch","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreadUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/threads/{thread_id}/messages":{"post":{"tags":["Threads","Threads"],"summary":"Add Message","operationId":"add_message_api_threads__thread_id__messages_post","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Threads","Threads"],"summary":"List Messages","operationId":"list_messages_api_threads__thread_id__messages_get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"before","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Before"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/threads/{thread_id}/docs":{"post":{"tags":["Threads","Threads"],"summary":"Set Docs","operationId":"set_docs_api_threads__thread_id__docs_post","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocsSet"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Threads","Threads"],"summary":"Get Docs","operationId":"get_docs_api_threads__thread_id__docs_get","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/threads/{thread_id}/export":{"post":{"tags":["Threads","Threads"],"summary":"Export Thread","operationId":"export_thread_api_threads__thread_id__export_post","parameters":[{"name":"thread_id","in":"path","required":true,"schema":{"type":"string","title":"Thread Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/memory/tend-garden":{"post":{"tags":["Memory & Knowledge","Memory & Knowledge"],"summary":"Déclenche la tâche de consolidation de la mémoire","description":"Lance manuellement le 'MemoryGardener' pour analyser les sessions récentes, consolider les concepts dans la base de connaissance et gérer le cycle de vie des souvenirs.","operationId":"tend_garden_endpoint_api_memory_tend_garden_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Tend Garden Endpoint Api Memory Tend Garden Post"}}}}}}},"/dev-auth.html":{"get":{"summary":"Get Dev Auth Html","description":"Sert la page GIS de debug pour récupérer un ID token (JWT).","operationId":"get_dev_auth_html_dev_auth_html_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/api/_dev-auth-diag":{"get":{"summary":"Dev Auth Diag","operationId":"dev_auth_diag_api__dev_auth_diag_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/_static-diag":{"get":{"tags":["Health"],"summary":" Static Diag","operationId":"_static_diag_api__static_diag_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_document_api_documents_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_documents_upload_post"},"DocsSet":{"properties":{"doc_ids":{"items":{"type":"integer"},"type":"array","title":"Doc Ids"},"mode":{"type":"string","pattern":"^(replace|append)$","title":"Mode","default":"replace"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight","default":1.0}},"type":"object","required":["doc_ids"],"title":"DocsSet"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MessageCreate":{"properties":{"role":{"type":"string","pattern":"^(user|assistant|system|note)$","title":"Role"},"content":{"type":"string","title":"Content"},"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tokens"},"meta":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta"}},"type":"object","required":["role","content"],"title":"MessageCreate"},"ThreadCreate":{"properties":{"type":{"type":"string","pattern":"^(chat|debate)$","title":"Type"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"meta":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta"}},"type":"object","required":["type"],"title":"ThreadCreate"},"ThreadUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"agent_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Agent Id"},"archived":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Archived"},"meta":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta"}},"type":"object","title":"ThreadUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}